requirements.txt
aiosqlite==0.21.0
anyio==4.7.0
asyncio==3.4.3
certifi==2024.12.14
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
python-dotenv==1.0.1
python-telegram-bot==21.9
sniffio==1.3.1
typing_extensions==4.12.2
